Update Raft Implementation

+++ raft-core/src/main/scala/com/grok/raft/core/internal/LogEntry.scala
@@ -3,7 +3,8 @@ package com.grok.raft.core.internal
 /** Represents a single entry in the Raft log.
   */
 trait LogEntry {
-  val term: Int
-  val index: Int
+  val term: Long
+  val index: Long
 }
-case class LogEntryGeneral[T](term: Int, index: Int, command: T) extends LogEntry
+
+case class LogEntryGeneral[T](term: Long, index: Long, command: T) extends LogEntry
